requirements.txt
mysqlclient
mypy
python-binance
python-dotenv
sqlalchemy
Twisted
